@@ -1,58 +1,58 @@
 {
-  "name": "ReactNativeFastImageExample",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
-    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
-    "start": "react-native start",
-    "test": "jest"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "preset": "react-native"
-  },
-  "dependencies": {
-    "@react-native-community/masked-view": "^0.1.7",
-    "@react-navigation/bottom-tabs": "^5.1.0",
-    "@react-navigation/native": "^5.0.8",
-    "@react-navigation/stack": "^5.1.0",
-    "react": "16.9.0",
-    "react-native": "0.61.5",
-    "react-native-fast-image": "^8.1.3",
-    "react-native-gesture-handler": "^1.6.0",
-    "react-native-image-picker": "^2.3.1",
-    "react-native-image-progress": "^1.1.1",
-    "react-native-reanimated": "^1.7.0",
-    "react-native-safe-area-context": "^0.7.3",
-    "react-native-screens": "^2.2.0",
-    "react-native-status-bar-height": "^2.1.0",
-    "react-native-vector-icons": "^6.6.0",
-    "react-timeout": "^1.1.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/runtime": "^7.6.2",
-    "@react-native-community/eslint-config": "^0.0.5",
-    "@types/jest": "^24.0.24",
-    "@types/react-native": "^0.60.25",
-    "@types/react-native-vector-icons": "^6.4.5",
-    "@types/react-test-renderer": "16.9.1",
-    "@typescript-eslint/eslint-plugin": "^2.12.0",
-    "@typescript-eslint/parser": "^2.12.0",
-    "babel-jest": "^24.9.0",
-    "eslint": "^6.5.1",
-    "jest": "^24.9.0",
-    "metro-react-native-babel-preset": "^0.56.0",
-    "react-test-renderer": "16.9.0",
-    "typescript": "^3.8.3"
-  }
+    "name": "reactnativefastimageexample",
+    "version": "0.0.1",
+    "private": true,
+    "scripts": {
+        "android": "react-native run-android",
+        "ios": "react-native run-ios",
+        "start": "react-native start",
+        "test": "jest",
+        "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
+    },
+    "jest": {
+        "moduleFileExtensions": [
+            "ts",
+            "tsx",
+            "js",
+            "jsx",
+            "json",
+            "node"
+        ],
+        "preset": "react-native"
+    },
+    "resolutions": {
+        "@types/react": "^17"
+    },
+    "dependencies": {
+        "@react-native-community/masked-view": "^0.1.7",
+        "@react-navigation/bottom-tabs": "^6.0.5",
+        "@react-navigation/native": "^6.0.2",
+        "@react-navigation/stack": "^6.0.7",
+        "react": "17.0.2",
+        "react-native": "0.65.1",
+        "react-native-fast-image": "^8.1.3",
+        "react-native-gesture-handler": "^1.6.0",
+        "react-native-image-picker": "^4.0.6",
+        "react-native-image-progress": "^1.1.1",
+        "react-native-reanimated": "^2.2.2",
+        "react-native-safe-area-context": "^3.3.2",
+        "react-native-screens": "^3.7.2",
+        "react-native-status-bar-height": "^2.1.0",
+        "react-native-vector-icons": "^8.1.0"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.12.9",
+        "@babel/runtime": "^7.12.5",
+        "@react-native-community/eslint-config": "^2.0.0",
+        "@types/jest": "^26.0.23",
+        "@types/react-native": "^0.65.0",
+        "@types/react-test-renderer": "^17.0.1",
+        "babel-jest": "^26.6.3",
+        "eslint": "^7.14.0",
+        "jest": "^26.6.3",
+        "metro-react-native-babel-preset": "^0.66.0",
+        "react-native-codegen": "^0.0.7",
+        "react-test-renderer": "17.0.2",
+        "typescript": "^3.8.3"
+    }
 }
