@@ -1,10 +1,10 @@
-import React, { Component } from 'react'
+import React, { useCallback, useMemo, useState } from 'react'
 import { StyleSheet, View } from 'react-native'
-import withCacheBust from './withCacheBust'
 import SectionFlex from './SectionFlex'
 import FastImage, { FastImageProps } from 'react-native-fast-image'
 import Section from './Section'
 import FeatureText from './FeatureText'
+import { useCacheBust } from './useCacheBust'
 
 const GIF_URL =
     'https://cdn-images-1.medium.com/max/1600/1*-CY5bU4OqiJRox7G00sftw.gif'
@@ -16,72 +16,59 @@ interface AutoSizingImageProps extends FastImageProps {
     style?: any
 }
 
-interface AutoSizingImageState {
-    height: number
-    width: number
-}
-
-class AutoSizingImage extends Component<
-    AutoSizingImageProps,
-    AutoSizingImageState
-> {
-    state = {
+const AutoSizingImage = (props: AutoSizingImageProps) => {
+    const [dimensions, setDimensions] = useState({
         height: 0,
         width: 0,
-    }
+    })
 
-    onLoad = (e: any) => {
-        const {
-            nativeEvent: { width, height },
-        } = e
-        this.setState({ width, height })
-        if (this.props.onLoad) {
-            this.props.onLoad(e)
+    const propsOnLoad = props.onLoad
+    const onLoad = useCallback(
+        (e: any) => {
+            const {
+                nativeEvent: { width, height },
+            } = e
+            setDimensions({ width, height })
+            if (propsOnLoad) {
+                propsOnLoad(e)
+            }
+        },
+        [propsOnLoad],
+    )
+
+    const height = useMemo(() => {
+        if (!dimensions.height) {
+            return props.defaultHeight === undefined ? 300 : props.defaultHeight
         }
-    }
-
-    getHeight = () => {
-        if (!this.state.height) {
-            return this.props.defaultHeight === undefined
-                ? 300
-                : this.props.defaultHeight
-        }
-        const ratio = this.state.height / this.state.width
-        const height = this.props.width * ratio
-        return height
-    }
-
-    render() {
-        const height = this.getHeight()
-        return (
-            <FastImage
-                {...this.props}
-                onLoad={this.onLoad}
-                style={[{ width: this.props.width, height }, this.props.style]}
-            />
-        )
-    }
+        const ratio = dimensions.height / dimensions.width
+        return props.width * ratio
+    }, [dimensions.height, dimensions.width, props.defaultHeight, props.width])
+    return (
+        <FastImage
+            {...props}
+            onLoad={onLoad}
+            style={[{ width: props.width, height }, props.style]}
+        />
+    )
 }
 
-interface AutoSizeExampleProps {
-    onPressReload: () => void
-    bust: boolean
+export const AutoSizeExample = () => {
+    const { bust, url } = useCacheBust(GIF_URL)
+    return (
+        <View>
+            <Section>
+                <FeatureText text="• AutoSize." />
+            </Section>
+            <SectionFlex onPress={bust}>
+                <AutoSizingImage
+                    style={styles.image}
+                    width={200}
+                    source={{ uri: url }}
+                />
+            </SectionFlex>
+        </View>
+    )
 }
-
-const AutoSizeExample = ({ onPressReload, bust }: AutoSizeExampleProps) => (
-    <View>
-        <Section>
-            <FeatureText text="• AutoSize." />
-        </Section>
-        <SectionFlex onPress={onPressReload}>
-            <AutoSizingImage
-                style={styles.image}
-                width={200}
-                source={{ uri: GIF_URL + bust }}
-            />
-        </SectionFlex>
-    </View>
-)
 
 const styles = StyleSheet.create({
     image: {
@@ -90,5 +77,3 @@ const styles = StyleSheet.create({
         flex: 0,
     },
 })
-
-export default withCacheBust(AutoSizeExample)
