@@ -1,9 +1,8 @@
 module.exports = {
     root: true,
     extends: '@react-native-community',
-    parser: '@typescript-eslint/parser',
-    plugins: ['@typescript-eslint'],
     rules: {
         semi: ['error', 'never'],
+        'react-native/no-inline-styles': 'off',
     },
 }
