@@ -10,7 +10,7 @@ module.exports = {
         getTransformOptions: async () => ({
             transform: {
                 experimentalImportSupport: false,
-                inlineRequires: false,
+                inlineRequires: true,
             },
         }),
     },
