@@ -1,6 +1,6 @@
 import React from 'react'
 import { Image } from 'react-native'
-import ImageGrid from './ImageGrid'
+import { ImageGrid } from './ImageGrid'
 
 const DefaultImageGrid = () => <ImageGrid ImageComponent={Image} />
 
