@@ -1,17 +1,17 @@
 import React from 'react'
 import { ScrollView, StatusBar, StyleSheet, Text, View } from 'react-native'
 import Section from './Section'
-import PriorityExample from './PriorityExample'
-import GifExample from './GifExample'
-import BorderRadiusExample from './BorderRadiusExample'
 import FeatureText from './FeatureText'
-import ProgressExample from './ProgressExample'
-import PreloadExample from './PreloadExample'
-import ResizeModeExample from './ResizeModeExample'
-import TintColorExample from './TintColorExample'
-import LocalImagesExample from './LocalImagesExample'
 import StatusBarUnderlay, { STATUS_BAR_HEIGHT } from './StatusBarUnderlay'
-import AutoSizeExample from './AutoSizeExample'
+import { PriorityExample } from './PriorityExample'
+import { GifExample } from './GifExample'
+import { BorderRadiusExample } from './BorderRadiusExample'
+import { ProgressExample } from './ProgressExample'
+import { PreloadExample } from './PreloadExample'
+import { ResizeModeExample } from './ResizeModeExample'
+import { TintColorExample } from './TintColorExample'
+import { LocalImagesExample } from './LocalImagesExample'
+import { AutoSizeExample } from './AutoSizeExample'
 
 const FastImageExample = () => (
     <View style={styles.container}>
