@@ -1,6 +1,6 @@
 import React from 'react'
 import FastImage from 'react-native-fast-image'
-import ImageGrid from './ImageGrid'
+import { ImageGrid } from './ImageGrid'
 
 const FastImageGrid = () => <ImageGrid ImageComponent={FastImage} />
 
