@@ -6,12 +6,11 @@ import {
     TouchableOpacity,
     ViewProps,
 } from 'react-native'
-import withCacheBust from './withCacheBust'
 import FastImage, { FastImageProps, Source } from 'react-native-fast-image'
 import Section from './Section'
 import FeatureText from './FeatureText'
 import FieldsBase64 from './images/fields'
-import ImagePicker from 'react-native-image-picker'
+import { launchImageLibrary } from 'react-native-image-picker'
 import BulletText from './BulletText'
 
 // @ts-ignore
@@ -22,15 +21,6 @@ import FieldsWebP from './images/fields.webp'
 import JellyfishGIF from './images/jellyfish.gif'
 // @ts-ignore
 import JellyfishWebP from './images/jellyfish.webp'
-
-const options = {
-    title: 'Select Avatar',
-    customButtons: [{ name: 'fb', title: 'Choose Photo from Facebook' }],
-    storageOptions: {
-        skipBackup: true,
-        path: 'images',
-    },
-}
 
 const Image = ({ source, ...p }: FastImageProps) => (
     <FastImage style={styles.imageSquare} source={source} {...p} />
@@ -60,18 +50,18 @@ class PhotoExample extends Component<{}, PhotoExampleState> {
     state: PhotoExampleState = {}
 
     pick = () => {
-        ImagePicker.showImagePicker(options, response => {
+        launchImageLibrary({ mediaType: 'photo' }, (response) => {
             if (response.didCancel) {
                 console.log('ImagePicker - User cancelled.')
-            } else if (response.error) {
-                console.log(`ImagePicker - Error ${response.error}.`)
-            } else if (response.customButton) {
-                console.log(`ImagePicker - Tapped ${response.customButton}`)
+            } else if (response.errorCode) {
+                console.log(`ImagePicker - Error ${response.errorMessage}.`)
             } else {
-                const uri = response.uri
-                this.setState({
-                    image: { uri: uri },
-                })
+                const uri = response?.assets?.[0]?.uri
+                if (uri) {
+                    this.setState({
+                        image: { uri: uri },
+                    })
+                }
             }
         })
     }
@@ -93,7 +83,7 @@ class PhotoExample extends Component<{}, PhotoExampleState> {
     }
 }
 
-const LocalImagesExample = () => (
+export const LocalImagesExample = () => (
     <View>
         <Section>
             <FeatureText>• Local images.</FeatureText>
@@ -142,5 +132,3 @@ const styles = StyleSheet.create({
         right: 0,
     },
 })
-
-export default withCacheBust(LocalImagesExample)
