@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { useState } from 'react'
 import { StyleSheet, View } from 'react-native'
 import SectionFlex from './SectionFlex'
 import FastImage from 'react-native-fast-image'
@@ -7,70 +7,50 @@ import FeatureText from './FeatureText'
 import Button from './Button'
 // @ts-ignore
 import { createImageProgress } from 'react-native-image-progress'
+import { useCacheBust } from './useCacheBust'
 
 const IMAGE_URL =
     'https://cdn-images-1.medium.com/max/1600/1*-CY5bU4OqiJRox7G00sftw.gif'
 
 const Image = createImageProgress(FastImage)
 
-interface PreloadExampleProps {}
+export const PreloadExample = () => {
+    const [show, setShow] = useState(false)
+    const { url, bust } = useCacheBust(IMAGE_URL)
 
-class PreloadExample extends Component<PreloadExampleProps> {
-    state = {
-        show: false,
-        url: IMAGE_URL,
+    const preload = () => {
+        FastImage.preload([{ uri: url }])
     }
 
-    bustCache = () => {
-        const key = Math.random().toString()
-        const bust = `?bust=${key}`
-        // Preload images. This can be called anywhere.
-        const url = IMAGE_URL + bust
-        this.setState({
-            url,
-            show: false,
-        })
-    }
-
-    preload = () => {
-        FastImage.preload([{ uri: this.state.url }])
-    }
-
-    showImage = () => {
-        this.setState({ show: true })
-    }
-
-    render() {
-        return (
-            <View>
-                <Section>
-                    <FeatureText text="• Preloading." />
-                    <FeatureText text="• Progress indication using react-native-image-progress." />
-                </Section>
-                <SectionFlex style={styles.section}>
-                    {this.state.show ? (
-                        <Image
-                            style={styles.image}
-                            source={{ uri: this.state.url }}
-                        />
-                    ) : (
-                        <View style={styles.image} />
-                    )}
-                    <View style={styles.buttons}>
-                        <View style={styles.buttonView}>
-                            <Button text="Bust" onPress={this.bustCache} />
-                        </View>
-                        <View style={styles.buttonView}>
-                            <Button text="Preload" onPress={this.preload} />
-                        </View>
-                        <View style={styles.buttonView}>
-                            <Button text="Render" onPress={this.showImage} />
-                        </View>
+    return (
+        <View>
+            <Section>
+                <FeatureText text="• Preloading." />
+                <FeatureText text="• Progress indication using react-native-image-progress." />
+            </Section>
+            <SectionFlex style={styles.section}>
+                {show ? (
+                    <Image style={styles.image} source={{ uri: url }} />
+                ) : (
+                    <View style={styles.image} />
+                )}
+                <View style={styles.buttons}>
+                    <View style={styles.buttonView}>
+                        <Button text="Bust" onPress={bust} />
                     </View>
-                </SectionFlex>
-            </View>
-        )
-    }
+                    <View style={styles.buttonView}>
+                        <Button text="Preload" onPress={preload} />
+                    </View>
+                    <View style={styles.buttonView}>
+                        <Button
+                            text={show ? 'Hide' : 'Show'}
+                            onPress={() => setShow((v) => !v)}
+                        />
+                    </View>
+                </View>
+            </SectionFlex>
+        </View>
+    )
 }
 
 const styles = StyleSheet.create({
@@ -92,5 +72,3 @@ const styles = StyleSheet.create({
         width: 100,
     },
 })
-
-export default PreloadExample
