@@ -1,10 +1,10 @@
 import React from 'react'
 import { PixelRatio, StyleSheet, View } from 'react-native'
-import withCacheBust from './withCacheBust'
 import FastImage from 'react-native-fast-image'
 import Section from './Section'
 import SectionFlex from './SectionFlex'
 import FeatureText from './FeatureText'
+import { useCacheBust } from './useCacheBust'
 
 const getImageUrl = (id: string, width: number, height: number) =>
     `https://source.unsplash.com/${id}/${width}x${height}`
@@ -16,41 +16,39 @@ const IMAGE_URLS = [
     getImageUrl('S7VCcp6KCKE', IMAGE_SIZE, IMAGE_SIZE),
 ]
 
-interface PriorityExampleProps {
-    onPressReload: () => void
-    bust: string
+export const PriorityExample = () => {
+    const { query, bust } = useCacheBust('')
+    return (
+        <View>
+            <Section>
+                <FeatureText text="• Prioritize images (low, normal, high)." />
+            </Section>
+            <SectionFlex onPress={bust}>
+                <FastImage
+                    style={styles.image}
+                    source={{
+                        uri: IMAGE_URLS[0] + query,
+                        priority: FastImage.priority.low,
+                    }}
+                />
+                <FastImage
+                    style={styles.image}
+                    source={{
+                        uri: IMAGE_URLS[1] + query,
+                        priority: FastImage.priority.normal,
+                    }}
+                />
+                <FastImage
+                    style={styles.image}
+                    source={{
+                        uri: IMAGE_URLS[2] + query,
+                        priority: FastImage.priority.high,
+                    }}
+                />
+            </SectionFlex>
+        </View>
+    )
 }
-
-const PriorityExample = ({ onPressReload, bust }: PriorityExampleProps) => (
-    <View>
-        <Section>
-            <FeatureText text="• Prioritize images (low, normal, high)." />
-        </Section>
-        <SectionFlex onPress={onPressReload}>
-            <FastImage
-                style={styles.image}
-                source={{
-                    uri: IMAGE_URLS[0] + bust,
-                    priority: FastImage.priority.low,
-                }}
-            />
-            <FastImage
-                style={styles.image}
-                source={{
-                    uri: IMAGE_URLS[1] + bust,
-                    priority: FastImage.priority.normal,
-                }}
-            />
-            <FastImage
-                style={styles.image}
-                source={{
-                    uri: IMAGE_URLS[2] + bust,
-                    priority: FastImage.priority.high,
-                }}
-            />
-        </SectionFlex>
-    </View>
-)
 
 const styles = StyleSheet.create({
     image: {
@@ -61,5 +59,3 @@ const styles = StyleSheet.create({
         marginVertical: 20,
     },
 })
-
-export default withCacheBust(PriorityExample)
