@@ -1,6 +1,5 @@
 import React from 'react'
 import { StyleSheet, View } from 'react-native'
-import withCacheBust from './withCacheBust'
 import FastImage from 'react-native-fast-image'
 import Section from './Section'
 import SectionFlex from './SectionFlex'
@@ -9,35 +8,33 @@ import FeatureText from './FeatureText'
 // @ts-ignore
 import LogoImage from './images/logo.png'
 
-interface TintColorExampleProps {
-    onPressReload: () => void
+export const TintColorExample = () => {
+    return (
+        <View>
+            <Section>
+                <FeatureText text="Images with tint color." />
+                <FeatureText text="All non-transparent pixels are changed to the color." />
+            </Section>
+            <SectionFlex>
+                <FastImage
+                    style={styles.image}
+                    tintColor={'green'}
+                    source={LogoImage}
+                />
+                <FastImage
+                    style={styles.image}
+                    tintColor={'#9324c3'}
+                    source={LogoImage}
+                />
+                <FastImage
+                    style={styles.image}
+                    tintColor={'rgba(0,0,0,0.5)'}
+                    source={LogoImage}
+                />
+            </SectionFlex>
+        </View>
+    )
 }
-
-const TintColorExample = ({ onPressReload }: TintColorExampleProps) => (
-    <View>
-        <Section>
-            <FeatureText text="Images with tint color." />
-            <FeatureText text="All non-transparent pixels are changed to the color." />
-        </Section>
-        <SectionFlex onPress={onPressReload}>
-            <FastImage
-                style={styles.image}
-                tintColor={'green'}
-                source={LogoImage}
-            />
-            <FastImage
-                style={styles.image}
-                tintColor={'#9324c3'}
-                source={LogoImage}
-            />
-            <FastImage
-                style={styles.image}
-                tintColor={'rgba(0,0,0,0.5)'}
-                source={LogoImage}
-            />
-        </SectionFlex>
-    </View>
-)
 
 const styles = StyleSheet.create({
     image: {
@@ -46,5 +43,3 @@ const styles = StyleSheet.create({
         margin: 10,
     },
 })
-
-export default withCacheBust(TintColorExample)
