@@ -1,5 +1,5 @@
 import React from 'react'
-import { YellowBox } from 'react-native'
+import { LogBox } from 'react-native'
 import { NavigationContainer } from '@react-navigation/native'
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
 import { Icon } from './Icon'
@@ -9,7 +9,7 @@ import DefaultImageGrid from './DefaultImageGrid'
 
 const Tab = createBottomTabNavigator()
 
-YellowBox.ignoreWarnings([
+LogBox.ignoreLogs([
     'Warning: isMounted(...) is deprecated',
     'Module RCTImageLoader',
 ])
@@ -17,12 +17,12 @@ YellowBox.ignoreWarnings([
 export default function App() {
     return (
         <NavigationContainer>
-            <Tab.Navigator>
+            <Tab.Navigator screenOptions={{ headerShown: false }}>
                 <Tab.Screen
                     name="FastImage Example"
                     component={FastImageExamples}
                     options={{
-                        tabBarIcon: props => (
+                        tabBarIcon: (props) => (
                             <Icon name="ios-information-circle" {...props} />
                         ),
                     }}
@@ -31,8 +31,8 @@ export default function App() {
                     name="Image Grid"
                     component={DefaultImageGrid}
                     options={{
-                        tabBarIcon: props => (
-                            <Icon name="ios-image" {...props} />
+                        tabBarIcon: (props) => (
+                            <Icon name="image-outline" {...props} />
                         ),
                     }}
                 />
@@ -40,8 +40,8 @@ export default function App() {
                     name="FastImage Grid"
                     component={FastImageGrid}
                     options={{
-                        tabBarIcon: props => (
-                            <Icon name="ios-photos" {...props} />
+                        tabBarIcon: (props) => (
+                            <Icon name="images-outline" {...props} />
                         ),
                     }}
                 />
